Add in schemas, clean up areas of pattern

+++ src/app/devices/model.rs
@@ -1,11 +1,13 @@
-#[derive(Debug)]
+use serde::Serialize;
+
+#[derive(Debug, Serialize)]
 pub struct Device {
-    pub id: i32,
+    pub id: i64,
     pub serial_number: String,
 }
 
 impl Device {
-    pub fn new(id: i32, serial_number: &str) -> Self {
+    pub fn new(id: i64, serial_number: &str) -> Self {
         Self {
             id,
             serial_number: serial_number.to_string(),
